Aplicando DIP e refatorando o codigo e tirando o acoplamento resultante da classe controle

+++ src/main/java/br/com/ucsal/olimpiadas/Participante.java
@@ -8,11 +8,11 @@ public class Participante {
 	
 	
 	
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
